keep cluster in module scope
