+++ tests/integration_tests/conftest.py
@@ -1,8 +1,6 @@
-from pathlib import Path
-
 import pytest
 
-from .network import setup_custom_ethermint, setup_ethermint, setup_geth
+from .network import setup_ethermint, setup_geth
 
 
 @pytest.fixture(scope="session")
@@ -12,45 +10,9 @@ def ethermint(tmp_path_factory):
 
 
 @pytest.fixture(scope="session")
-def ethermint_long_timeout_commit(tmp_path_factory):
-    path = tmp_path_factory.mktemp("long_timeout_commit")
-    yield from setup_ethermint(path, 26200, True)
-
-
-@pytest.fixture(scope="session")
-def ethermint_indexer(tmp_path_factory):
-    path = tmp_path_factory.mktemp("indexer")
-    yield from setup_custom_ethermint(
-        path, 26660, Path(__file__).parent / "configs/enable-indexer.jsonnet"
-    )
-
-
-@pytest.fixture(scope="session")
 def geth(tmp_path_factory):
     path = tmp_path_factory.mktemp("geth")
     yield from setup_geth(path, 8545)
-
-
-@pytest.fixture(
-    scope="session", params=["ethermint", "geth", "ethermint-ws", "enable-indexer"]
-)
-def cluster(request, ethermint_long_timeout_commit, ethermint_indexer, geth):
-    """
-    run on both ethermint and geth
-    """
-    provider = request.param
-    if provider == "ethermint":
-        yield ethermint_long_timeout_commit
-    elif provider == "geth":
-        yield geth
-    elif provider == "ethermint-ws":
-        ethermint_ws = ethermint_long_timeout_commit.copy()
-        ethermint_ws.use_websocket()
-        yield ethermint_ws
-    elif provider == "enable-indexer":
-        yield ethermint_indexer
-    else:
-        raise NotImplementedError
 
 
 @pytest.fixture(
